chore(refactor): break out components of application state
