+++ types/Team.ts
@@ -1,0 +1,4 @@
+export type Team = "radiant" | "dire" | "bans";
+// map to https://docs.opendota.com/#tag/scenarios%2Fpaths%2F~1scenarios~1laneRoles%2Fget
+export type Role = "carry" | "mid" | "offlaner" | "softSupport" | "hardSupport";
+export type Lane = "safelane" | "middle" | "offlane" | "jungle";
